Refactor address intiailization, put it inside server state
