+++ server/status.rs
@@ -9,7 +9,22 @@ use typedb_protocol;
 
 #[derive(Clone, Copy, Debug)]
 pub struct LocalServerStatus {
-    pub address: SocketAddr,
+    grpc_address: SocketAddr,
+    http_address: Option<SocketAddr>,
+}
+
+impl LocalServerStatus {
+    pub(crate) fn new(grpc_address: SocketAddr, http_address: Option<SocketAddr>) -> Self {
+        Self { grpc_address, http_address }
+    }
+
+    pub(crate) fn grpc_address(&self) -> SocketAddr {
+        self.grpc_address
+    }
+
+    pub(crate) fn http_address(&self) -> Option<SocketAddr> {
+        self.http_address
+    }
 }
 
 pub trait ServerStatus: Debug {
@@ -18,6 +33,6 @@ pub trait ServerStatus: Debug {
 
 impl ServerStatus for LocalServerStatus {
     fn to_proto(&self) -> typedb_protocol::Server {
-        typedb_protocol::Server { address: self.address.to_string(), replica_status: None }
+        typedb_protocol::Server { address: self.grpc_address.to_string(), replica_status: None }
     }
 }
